ROX-9653: Cluster Init Bundles in Postgres
