+++ central/clusterinit/backend/backend_impl.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/gogo/protobuf/types"
 	"github.com/pkg/errors"
+	"github.com/stackrox/rox/central/clusterinit/backend/access"
 	"github.com/stackrox/rox/central/clusterinit/backend/certificate"
 	"github.com/stackrox/rox/central/clusterinit/store"
 	"github.com/stackrox/rox/central/clusters"
@@ -24,7 +25,7 @@ type backendImpl struct {
 }
 
 func (b *backendImpl) GetAll(ctx context.Context) ([]*storage.InitBundleMeta, error) {
-	if err := CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
+	if err := access.CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
 		return nil, err
 	}
 
@@ -74,7 +75,7 @@ func extractExpiryDate(certBundle clusters.CertBundle) (*types.Timestamp, error)
 }
 
 func (b *backendImpl) Issue(ctx context.Context, name string) (*InitBundleWithMeta, error) {
-	if err := CheckAccess(ctx, storage.Access_READ_WRITE_ACCESS); err != nil {
+	if err := access.CheckAccess(ctx, storage.Access_READ_WRITE_ACCESS); err != nil {
 		return nil, err
 	}
 
@@ -120,7 +121,7 @@ func (b *backendImpl) Issue(ctx context.Context, name string) (*InitBundleWithMe
 }
 
 func (b *backendImpl) GetCAConfig(ctx context.Context) (*CAConfig, error) {
-	if err := CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
+	if err := access.CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
 		return nil, err
 	}
 
@@ -135,7 +136,7 @@ func (b *backendImpl) GetCAConfig(ctx context.Context) (*CAConfig, error) {
 }
 
 func (b *backendImpl) Revoke(ctx context.Context, id string) error {
-	if err := CheckAccess(ctx, storage.Access_READ_WRITE_ACCESS); err != nil {
+	if err := access.CheckAccess(ctx, storage.Access_READ_WRITE_ACCESS); err != nil {
 		return err
 	}
 
@@ -147,7 +148,7 @@ func (b *backendImpl) Revoke(ctx context.Context, id string) error {
 }
 
 func (b *backendImpl) CheckRevoked(ctx context.Context, id string) error {
-	if err := CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
+	if err := access.CheckAccess(ctx, storage.Access_READ_ACCESS); err != nil {
 		return err
 	}
