+++ central/clusterinit/backend/backend_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/stackrox/rox/central/clusterinit/backend/certificate/mocks"
+	"github.com/stackrox/rox/central/clusterinit/store"
 	rocksdbStore "github.com/stackrox/rox/central/clusterinit/store/rocksdb"
 	"github.com/stackrox/rox/central/clusters"
 	"github.com/stackrox/rox/central/role/resources"
@@ -120,9 +121,9 @@ func (s *clusterInitBackendTestSuite) SetupTest() {
 	s.mockCtrl = gomock.NewController(s.T())
 	m := mocks.NewMockProvider(s.mockCtrl)
 	s.rocksDB = rocksdbtest.RocksDBForT(s.T())
-	store, err := rocksdbStore.NewStore(s.rocksDB)
+	rocksdbStore, err := rocksdbStore.New(s.rocksDB)
 	s.Require().NoError(err)
-	s.backend = newBackend(store, m)
+	s.backend = newBackend(store.NewStore(rocksdbStore), m)
 	s.ctx = sac.WithGlobalAccessScopeChecker(context.Background(), sac.AllowAllAccessScopeChecker())
 	s.certProvider = m
